FAC-534: Move in adapter to attribute package
