+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/attribute/CreateAssessmentAttributeAiReportRequestDto.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.adapter.in.rest.assessment;
+package org.flickit.assessment.core.adapter.in.rest.attribute;
 
 public record CreateAssessmentAttributeAiReportRequestDto(String fileLink) {
 }
