+++ flickit-assessment-core/src/main/java/org/flickit/assessment/core/adapter/in/rest/attribute/CreateAssessmentAttributeAiReportResponseDto.java
@@ -1,4 +1,4 @@
-package org.flickit.assessment.core.adapter.in.rest.assessment;
+package org.flickit.assessment.core.adapter.in.rest.attribute;
 
 public record CreateAssessmentAttributeAiReportResponseDto(String content) {
 }
